+++ templates/pilet-template-ng-standalone/src/versions.ts
@@ -1,0 +1,14 @@
+// Mapping of Angular to TypeScript versions
+export const tsVersions = {
+  17: '~5.2',
+};
+
+// Mapping of Angular to RxJs versions
+export const rxjsVersions = {
+  17: '^7.4',
+};
+
+// Mapping of Angular to Zone.js versions
+export const zoneVersions = {
+  17: '0.14.0',
+};